context as a method
This change gives a way to define own context, thus, it gets passed
to a resource.

The idea was to define a `context` method which could be redefined
by children. But, it looks like a child's context method gets redefined
but the helper method. In general, any method listed in `Grape::JSONAPI::Helpers`
cannot be redefined in children.